hands/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
asyncclick==8.3.0.4
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mashumaro==3.17
multidict==6.6.4
PlexAPI==4.17.1
propcache==0.4.0
pycparser==2.23
PyMySQL==1.1.2
pyspnego==0.12.0
python-dotenv==1.1.1
python-kasa==0.10.2
pywinrm==0.5.0
requests==2.32.0
requests_ntlm==1.3.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xmltodict==1.0.2
yarl==1.21.0
